Rewrote CMMapView as CMMapView2.
This does away with using CPWebView - it has an iframe in it, and this caused issues if you dragged out the map. Although a nice container for generic bits of the web, it's overkill for us, so this is a much lighter weight implementation.

Note that you now need to include the following in your root html file:

	<script type="text/javascript" src="http://tile.cloudmade.com/wml/latest/web-maps-lite.js"></script>